fix(ci): split two files back under the 800-line gate the red was hiding
`make check-file-sizes` is CI's step 15 in the `test` job. That job has been dying at
step 11 since 2026-09-02, so steps 12-60 have been `skipped` for ~24 hours — 45 gates
dark. Measured across the boundary: at the last green `test` job (`7668ed9df`, 13:18)
`backend_gcp.go` was 788 lines and `inbox.go` 773, and step 15 read `success`. They are
811 and 850 now. Both crossed the limit inside the window where nothing could see them.
So landing the fmt A/B fix alone would not have made dev green; it would have moved the
failure from step 11 to step 15.

PURE MOVES, same package each, following each package's existing multi-file convention
(`backend_controlplane_*.go`, `inbox_routing.go`):

  internal/observatory/backend_gcp.go  811 -> 657  (+ backend_gcp_chains.go, 160)
  internal/messaging/inbox.go          850 -> 749  (+ inbox_flow.go,          102)

Purity measured, not asserted: of the 116 and 91 non-blank lines removed from the two
originals, **0** are absent from the corresponding new file, and **0** lines were added
to either original. The new files carry 5 and 1 extra non-blank lines — the package
clause and imports. Controls on that instrument fire in both directions (the package
clause is found; a fabricated marker is not). A first run of this check reported 91
lines missing and was an INSTRUMENT failure of mine, not a finding: the header-skip
heuristic consumed the whole file, which the `body_nonblank=0` reading exposed.

Gates re-run by the controller OUTSIDE the executor sandbox, because the executor
correctly labelled its own `go test` rc=1 UNINFORMATIVE UNDER SANDBOX — a loopback bind
denial in `TestHub_WebSocketIntegration`, not a regression. All rc=0: scoped
`go build ./internal/... ./cmd/ailang`, `go vet`, `go test` over observatory + messaging
+ eval_analysis, `make check-file-sizes` (rc=2 -> 0), `fmt-check`, `check-boundaries`,
`check-referenced-paths`, `make test-launchd-drivers` (rc=2 -> 0, 43 probe arms).
`go build ./...` stays rc=1 on `cmd/wasm`, which is red at base and unrelated.

Reported at #972.
